masc: update pending balance display
